--- a/core/src/main/java/kafka/log/remote/RemoteLogOffsetReader.java
+++ b/core/src/main/java/kafka/log/remote/RemoteLogOffsetReader.java
@@ -19,6 +19,7 @@ package kafka.log.remote;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.record.FileRecords;
 import org.apache.kafka.storage.internals.epoch.LeaderEpochFileCache;
+import org.apache.kafka.storage.internals.log.AsyncOffsetReader;
 import org.apache.kafka.storage.internals.log.OffsetResultHolder;
 
 import org.slf4j.Logger;
@@ -27,7 +28,6 @@ import org.slf4j.LoggerFactory;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 public class RemoteLogOffsetReader implements Callable<Void> {
     private static final Logger LOGGER = LoggerFactory.getLogger(RemoteLogOffsetReader.class);
@@ -36,7 +36,7 @@ public class RemoteLogOffsetReader implements Callable<Void> {
     private final long timestamp;
     private final long startingOffset;
     private final LeaderEpochFileCache leaderEpochCache;
-    private final Supplier<Optional<FileRecords.TimestampAndOffset>> searchInLocalLog;
+    private final AsyncOffsetReader.TimestampAndOffsetSupplier searchInLocalLog;
     private final Consumer<OffsetResultHolder.FileRecordsOrError> callback;
 
     public RemoteLogOffsetReader(RemoteLogManager rlm,
@@ -44,7 +44,7 @@ public class RemoteLogOffsetReader implements Callable<Void> {
                                  long timestamp,
                                  long startingOffset,
                                  LeaderEpochFileCache leaderEpochCache,
-                                 Supplier<Optional<FileRecords.TimestampAndOffset>> searchInLocalLog,
+                                 AsyncOffsetReader.TimestampAndOffsetSupplier searchInLocalLog,
                                  Consumer<OffsetResultHolder.FileRecordsOrError> callback) {
         this.rlm = rlm;
         this.tp = tp;
@@ -60,8 +60,11 @@ public class RemoteLogOffsetReader implements Callable<Void> {
         OffsetResultHolder.FileRecordsOrError result;
         try {
             // If it is not found in remote storage, then search in the local storage starting with local log start offset.
-            Optional<FileRecords.TimestampAndOffset> timestampAndOffsetOpt = 
-                    rlm.findOffsetByTimestamp(tp, timestamp, startingOffset, leaderEpochCache).or(searchInLocalLog);
+            Optional<FileRecords.TimestampAndOffset> timestampAndOffsetOpt =
+                    rlm.findOffsetByTimestamp(tp, timestamp, startingOffset, leaderEpochCache);
+            if (timestampAndOffsetOpt.isEmpty()) {
+                timestampAndOffsetOpt = searchInLocalLog.get();
+            }
             result = new OffsetResultHolder.FileRecordsOrError(Optional.empty(), timestampAndOffsetOpt);
         } catch (Exception e) {
             // NOTE: All the exceptions from the secondary storage are caught instead of only the KafkaException.