--- a/core/src/main/scala/kafka/server/AbstractFetcherThread.scala
+++ b/core/src/main/scala/kafka/server/AbstractFetcherThread.scala
@@ -643,7 +643,6 @@ abstract class AbstractFetcherThread(name: String,
     }
   }
 
-
   /**
    * Handle a partition whose offset is out of range and return a new fetch offset.
    */