… checkstyle/suppressions.xml
… core/src/main/scala/kafka/log/LogManager.scala
… core/src/main/scala/kafka/log/UnifiedLog.scala
… core/src/main/scala/kafka/raft/KafkaMetadataLog.scala
… core/src/main/scala/kafka/server/ConfigHandler.scala
… core/src/main/scala/kafka/server/LocalLeaderEndPoint.scala
… core/src/main/scala/kafka/server/RemoteLeaderEndPoint.scala
… core/src/main/scala/kafka/server/ReplicaAlterLogDirsThread.scala
… core/src/main/scala/kafka/server/ReplicaFetcherThread.scala
… core/src/main/scala/kafka/server/ReplicaManager.scala
… core/src/main/scala/kafka/tools/DumpLogSegments.scala
… core/src/test/java/kafka/admin/DeleteTopicTest.java
… core/src/test/java/kafka/log/remote/RemoteLogManagerTest.java
… core/src/test/java/kafka/server/LogManagerIntegrationTest.java
… core/src/test/scala/integration/kafka/api/ConsumerWithLegacyMessageFormatIntegrationTest.scala
… core/src/test/scala/integration/kafka/api/GroupCoordinatorIntegrationTest.scala
… core/src/test/scala/integration/kafka/server/DynamicBrokerReconfigurationTest.scala
… core/src/test/scala/integration/kafka/server/KRaftClusterTest.scala
… core/src/test/scala/kafka/raft/KafkaMetadataLogTest.scala
… core/src/test/scala/kafka/server/LocalLeaderEndPointTest.scala
… core/src/test/scala/kafka/server/RemoteLeaderEndPointTest.scala
… core/src/test/scala/unit/kafka/cluster/PartitionLockTest.scala
… core/src/test/scala/unit/kafka/cluster/PartitionTest.scala
… core/src/test/scala/unit/kafka/coordinator/AbstractCoordinatorConcurrencyTest.scala
… core/src/test/scala/unit/kafka/coordinator/group/CoordinatorLoaderImplTest.scala
… core/src/test/scala/unit/kafka/coordinator/group/GroupMetadataManagerTest.scala
… core/src/test/scala/unit/kafka/coordinator/transaction/TransactionCoordinatorConcurrencyTest.scala
… core/src/test/scala/unit/kafka/coordinator/transaction/TransactionStateManagerTest.scala
… core/src/test/scala/unit/kafka/log/AbstractLogCleanerIntegrationTest.scala
… core/src/test/scala/unit/kafka/log/BrokerCompressionTest.scala