@mimaison 2025-03-05
The CI failure is caused by an issue in trunk, there's a PR to fix it: https://github.com/apache/kafka/pull/19115

I'll rekick the CI, once #19115 is merged.