@mimaison 2025-03-08
Rebased on trunk and ported https://github.com/apache/kafka/commit/40db001588047201a406ebe969d1d7d2d5eefd57