@junrao 2025-03-10
Also, could you run some more perf tests like kafka-producer-perf-test.sh and kafka-consumer-perf-test.sh to make sure there is no regression?