@mimaison 2025-03-10
> Also, could you run some more perf tests like kafka-producer-perf-test.sh and kafka-consumer-perf-test.sh to make sure there is no regression?

Yes once we're happy with the rewrite I'll run a few perf tests. Unfortunately I've not got access to a dedicated performance environment but I'll run tests on my laptop. Also we'll be early enough in the 4.1.0 release cycle to get feedback from the community test it before it releases.